requirements.txt
Flask==1.1.2
cloudinary==1.25.0
pymongo[tls,srv]>=3.11.4
python-dotenv==0.17.1
bcrypt==3.1.2
sendgrid==6.7.0
gunicorn==19.9.0